backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2020.1
sqlparse==0.3.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
